fix bugs

+++ src/v1/controllers/habitController.js
@@ -6,7 +6,7 @@ exports.createHabit = catchAsync(async (req, res, next) => {
   const createdHabit = await userService.createHabit(req.user, habitData);
   res.status(201).json({
     status: "OK",
-    data: createdHabit,
+    data: { hahit: createdHabit },
   });
 });
 exports.getHabits = (req, res, next) => {
@@ -14,13 +14,17 @@ exports.getHabits = (req, res, next) => {
   const habits = userService.getHabits(user);
   res.status(200).json({
     status: "OK",
-    data: habits,
+    data: { habits },
   });
 };
 
 exports.deleteHabit = catchAsync(async (req, res, next) => {
-  const { id: userId } = req.user;
+  const {
+    user: { id: userId },
+    user,
+  } = req;
   const habitId = req.params.habitId;
+  await userService.verifyHabitOwnership(user, habitId);
   await userService.deleteHabit(userId, habitId);
   res.status(204).json({
     status: "OK",
